+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/upstream/CmcdConfiguration.java
@@ -60,7 +60,10 @@ public final class CmcdConfiguration {
     KEY_BUFFER_LENGTH,
     KEY_CONTENT_ID,
     KEY_SESSION_ID,
-    KEY_MAXIMUM_REQUESTED_BITRATE
+    KEY_MAXIMUM_REQUESTED_BITRATE,
+    KEY_STREAMING_FORMAT,
+    KEY_STREAM_TYPE,
+    KEY_VERSION
   })
   @Documented
   @Target(TYPE_USE)
@@ -78,6 +81,9 @@ public final class CmcdConfiguration {
   public static final String KEY_CONTENT_ID = "cid";
   public static final String KEY_SESSION_ID = "sid";
   public static final String KEY_MAXIMUM_REQUESTED_BITRATE = "rtp";
+  public static final String KEY_STREAMING_FORMAT = "sf";
+  public static final String KEY_STREAM_TYPE = "st";
+  public static final String KEY_VERSION = "v";
 
   /**
    * Factory for {@link CmcdConfiguration} instances.
@@ -201,7 +207,7 @@ public final class CmcdConfiguration {
   }
 
   /**
-   * Whether logging bitrate is allowed based on the {@linkplain RequestConfig request
+   * Returns whether logging bitrate is allowed based on the {@linkplain RequestConfig request
    * configuration}.
    */
   public boolean isBitrateLoggingAllowed() {
@@ -209,7 +215,7 @@ public final class CmcdConfiguration {
   }
 
   /**
-   * Whether logging buffer length is allowed based on the {@linkplain RequestConfig request
+   * Returns whether logging buffer length is allowed based on the {@linkplain RequestConfig request
    * configuration}.
    */
   public boolean isBufferLengthLoggingAllowed() {
@@ -217,7 +223,7 @@ public final class CmcdConfiguration {
   }
 
   /**
-   * Whether logging content ID is allowed based on the {@linkplain RequestConfig request
+   * Returns whether logging content ID is allowed based on the {@linkplain RequestConfig request
    * configuration}.
    */
   public boolean isContentIdLoggingAllowed() {
@@ -225,7 +231,7 @@ public final class CmcdConfiguration {
   }
 
   /**
-   * Whether logging session ID is allowed based on the {@linkplain RequestConfig request
+   * Returns whether logging session ID is allowed based on the {@linkplain RequestConfig request
    * configuration}.
    */
   public boolean isSessionIdLoggingAllowed() {
@@ -233,10 +239,26 @@ public final class CmcdConfiguration {
   }
 
   /**
-   * Whether logging maximum requested throughput is allowed based on the {@linkplain RequestConfig
-   * request configuration}.
+   * Returns whether logging maximum requested throughput is allowed based on the {@linkplain
+   * RequestConfig request configuration}.
    */
   public boolean isMaximumRequestThroughputLoggingAllowed() {
     return requestConfig.isKeyAllowed(KEY_MAXIMUM_REQUESTED_BITRATE);
+  }
+
+  /**
+   * Returns whether logging streaming format is allowed based on the {@linkplain RequestConfig
+   * request configuration}.
+   */
+  public boolean isStreamingFormatLoggingAllowed() {
+    return requestConfig.isKeyAllowed(KEY_STREAMING_FORMAT);
+  }
+
+  /**
+   * Returns whether logging stream type is allowed based on the {@linkplain RequestConfig request
+   * configuration}.
+   */
+  public boolean isStreamTypeLoggingAllowed() {
+    return requestConfig.isKeyAllowed(KEY_STREAM_TYPE);
   }
 }