+++ libraries/exoplayer_dash/src/main/java/androidx/media3/exoplayer/dash/DefaultDashChunkSource.java
@@ -374,7 +374,12 @@ public class DefaultDashChunkSource implements DashChunkSource {
     CmcdLog cmcdLog =
         cmcdConfiguration == null
             ? null
-            : CmcdLog.createInstance(cmcdConfiguration, trackSelection, bufferedDurationUs);
+            : CmcdLog.createInstance(
+                cmcdConfiguration,
+                trackSelection,
+                bufferedDurationUs,
+                CmcdLog.STREAMING_FORMAT_DASH,
+                manifest.dynamic);
 
     RepresentationHolder representationHolder =
         updateSelectedBaseUrl(trackSelection.getSelectedIndex());