+++ libraries/exoplayer_dash/src/test/java/androidx/media3/exoplayer/dash/DefaultDashChunkSourceTest.java
@@ -318,7 +318,7 @@ public class DefaultDashChunkSourceTest {
             "CMCD-Request",
             "bl=0",
             "CMCD-Session",
-            "cid=\"mediaId\",sid=\"" + cmcdConfiguration.sessionId + "\"");
+            "cid=\"mediaId\",sid=\"" + cmcdConfiguration.sessionId + "\",sf=d,st=v");
   }
 
   @Test
@@ -363,7 +363,7 @@ public class DefaultDashChunkSourceTest {
             "CMCD-Request",
             "bl=0",
             "CMCD-Session",
-            "cid=\"mediaIdcontentIdSuffix\"",
+            "cid=\"mediaIdcontentIdSuffix\",sf=d,st=v",
             "CMCD-Status",
             "rtp=3500");
   }
@@ -409,7 +409,7 @@ public class DefaultDashChunkSourceTest {
             "CMCD-Request",
             "bl=0,key2=\"stringValue\"",
             "CMCD-Session",
-            "cid=\"mediaId\",sid=\"" + cmcdConfiguration.sessionId + "\",key3=1",
+            "cid=\"mediaId\",sid=\"" + cmcdConfiguration.sessionId + "\",sf=d,st=v,key3=1",
             "CMCD-Status",
             "key4=5.0");
   }