+++ libraries/exoplayer_smoothstreaming/src/main/java/androidx/media3/exoplayer/smoothstreaming/DefaultSsChunkSource.java
@@ -284,7 +284,12 @@ public class DefaultSsChunkSource implements SsChunkSource {
     CmcdLog cmcdLog =
         cmcdConfiguration == null
             ? null
-            : CmcdLog.createInstance(cmcdConfiguration, trackSelection, bufferedDurationUs);
+            : CmcdLog.createInstance(
+                cmcdConfiguration,
+                trackSelection,
+                bufferedDurationUs,
+                CmcdLog.STREAMING_FORMAT_SS,
+                manifest.isLive);
 
     out.chunk =
         newMediaChunk(