+++ RELEASENOTES.md
@@ -32,6 +32,8 @@
     *   Allow `MediaItem` updates for all `MediaSource` classes provided by the
         library via `Player.replaceMediaItem(s)`
         (([#33](https://github.com/androidx/media/issues/33)),([#9978](https://github.com/google/ExoPlayer/issues/9978))).
+    *   Add fields streaming format (sf), stream type (st) and version (v) to
+        Common Media Client Data (CMCD) logging.
 *   Transformer:
     *   Parse EXIF rotation data for image inputs.
     *   Remove `TransformationRequest.HdrMode` annotation type and its